+++ spec/lib/html2rss/selectors/extractors/static_spec.rb
@@ -3,7 +3,7 @@
 RSpec.describe Html2rss::Selectors::Extractors::Static do
   subject { described_class.new(nil, options).get }
 
-  let(:options) { instance_double(Struct::StaticOptions, static: 'Foobar') }
+  let(:options) { described_class::Options.new(static: 'Foobar') }
 
   it { is_expected.to eq 'Foobar' }
 end