+++ spec/lib/html2rss/selectors/post_processors/parse_uri_spec.rb
@@ -7,7 +7,7 @@ RSpec.describe Html2rss::Selectors::PostProcessors::ParseUri do
 
   let(:context) do
     Html2rss::Selectors::Context.new(
-      config: { channel: { url: 'http://example.com' } }
+      channel: { url: 'http://example.com' }
     )
   end